refactor: move InvoiceStatus and QuoteStatus enums to accounting app
The InvoiceStatus and QuoteStatus enums were located in the job app,
but they are actually used in the accounting app. This commit moves
them to the accounting app to improve the project structure and
maintainability.
