+++ apps/workflow/views/xero/xero_quote_manager.py
@@ -16,7 +16,7 @@ from .xero_helpers import (
 
 # Import models
 from apps.accounting.models import Quote
-from apps.job.enums import QuoteStatus
+from apps.accounting.enums import QuoteStatus
 from xero_python.accounting.models import LineItem, Quote as XeroQuote
 from xero_python.exceptions import (
     AccountingBadRequestException,
